Update DistanceSensor.c
This file contains the C code for the distance sensor whose Trigger Pin is connected to Pin PE2 and Echo Pin is connected to Pin PF4 of the Launchpad. 

A systick timer is initialized which triggers the Distance sensor every 50ms and updates the value of the variable distance2. You can comment out the Initialization of the Systick Timer and trigger the sensor manually in your own code.

Just connect the Distance sensor to these two pins and then Include the header file of this driver named as "distance_sensor.h" in your main program.

Call the function getDistance() to get the most recent distance value from the sensor in units of inches..